@@ -1,15 +1,15 @@
 name: notes
 description: "A new Flutter project."
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
-version: 1.9.2
+version: 2.0.0
 environment:
-  sdk: '>=3.3.4 <4.0.0'
+  sdk: '>=3.4.0 <4.0.0'
 dependencies:
   flutter_localizations:
     sdk: flutter
   flutter:
     sdk: flutter
-  intl: any
+  intl: ^0.19.0
   path: ^1.9.0
   hive: ^2.2.3
   get_it: ^7.7.0
